clear login auth from API response, apply to response cookies
