+++ src/apiProxy/api-proxy.js
@@ -5,6 +5,7 @@ const externalRequest$ = Rx.Observable.bindNodeCallback(externalRequest);
 
 import * as apiConfigCreators from './apiConfigCreators';
 import { duotoneRef } from '../util/duotone';
+import { applyAuth } from '../util/authUtils';
 
 const parseResponseFlags = ({ headers }) =>
 	(headers['x-meetup-flags'] || '')
@@ -250,6 +251,26 @@ export const apiResponseDuotoneSetter = duotoneUrls => {
 	};
 };
 
+const parseLoginAuth = (request, query) => response => {
+	if (query.type === 'login' && request.authorize) {
+		const {
+			oauth_token,
+			refresh_token,
+			expires_in,
+			member,
+		} = response.value;
+		applyAuth(request, request.authorize.reply)({
+			oauth_token,
+			refresh_token,
+			expires_in
+		});
+		return {
+			member
+		};
+	}
+	return response;
+};
+
 const MOCK_RESPONSE_OK = {  // minimal representation of http.IncomingMessage
 	statusCode: 200,
 	statusMessage: 'OK',
@@ -284,12 +305,13 @@ export const makeApiRequest$ = (request, API_TIMEOUT, duotoneUrls) => {
 			makeMockRequest(query.mockResponse) :
 			makeExternalApiRequest(request, API_TIMEOUT);
 
+
 		request.log(['api', 'info'], `REST API request: ${requestOpts.url}`);
 		return request$(requestOpts)
-			.map(parseApiResponse)             // parse into plain object
-			.catch(error => Rx.Observable.of({ error: error.message }))
-			.map(apiResponseToQueryResponse(query))    // convert apiResponse to app-ready queryResponse
-			.map(setApiResponseDuotones);        // special duotone prop
+			.map(parseApiResponse)                   // parse into plain object
+			.map(parseLoginAuth(request, query))     // login has oauth secrets - special case
+			.map(apiResponseToQueryResponse(query))  // convert apiResponse to app-ready queryResponse
+			.map(setApiResponseDuotones);            // special duotone prop
 	};
 };
